update table and api with defer
