+++ deepfence_frontend/apps/dashboard/src/features/registries/components/scan-forms/MalwareScanForm.tsx
@@ -1,0 +1,3 @@
+import { MalwareScanConfigureForm } from '@/components/registries/MalwareScanConfigureForm';
+
+export const MalwareScanForm = () => <MalwareScanConfigureForm />;
