+++ deepfence_frontend/apps/dashboard/src/features/registries/components/scan-forms/SecretScanForm.tsx
@@ -1,0 +1,3 @@
+import { SecretScanConfigureForm } from '@/components/registries/SecretScanConfigureForm';
+
+export const SecretScanForm = () => <SecretScanConfigureForm />;
