+++ deepfence_frontend/apps/dashboard/src/features/registries/components/scan-forms/VulnerabilityScanForms.tsx
@@ -1,0 +1,61 @@
+import { useState } from 'react';
+import { Form } from 'react-router-dom';
+import { Checkbox, Radio, TextInput } from 'ui-components';
+
+import { VulnerabilityScanConfigureForm } from '@/components/registries/VulnerabilityScanConfigureForm';
+
+export const VulnerabilityScanForm = () => {
+  const [priorityScan, setPriorityScan] = useState(false);
+  const [autoCheckandScan, setAutoCheckandScan] = useState(false);
+  return (
+    <Form className="flex flex-col">
+      <VulnerabilityScanConfigureForm
+        loading={false}
+        data={{
+          urlIds: [''],
+          urlType: '',
+        }}
+      />
+      <h6 className={'text-md font-medium dark:text-white mt-6'}>Advanced Options</h6>
+      <div className="mt-4 flex flex-col gap-y-6">
+        <Checkbox
+          name="priorityScan"
+          label="Priority Scan"
+          checked={priorityScan}
+          onCheckedChange={(checked: boolean) => {
+            setPriorityScan(checked);
+          }}
+        />
+        <Radio
+          name="scanTag"
+          defaultChecked
+          defaultValue="last"
+          options={[
+            { label: 'Scan last pushed tag', value: 'recent' },
+            { label: 'Scan by "latest" tag', value: 'last' },
+            { label: 'Scan all image tags', value: 'all' },
+          ]}
+          onValueChange={(value) => {
+            console.log('value', value);
+          }}
+        />
+        <TextInput
+          className="min-[200px] max-w-xs"
+          label="Scan interval in days (optional)"
+          type={'text'}
+          sizing="sm"
+          name="scanInterval"
+          placeholder=""
+        />
+        <Checkbox
+          name="scanEveryday"
+          label="Check and scan for new images every day"
+          checked={autoCheckandScan}
+          onCheckedChange={(checked: boolean) => {
+            setAutoCheckandScan(checked);
+          }}
+        />
+      </div>
+    </Form>
+  );
+};
